Include registration of salesforce xml namespaces when receiving soap xml
ElementTree is interesting in that it allows for capabilities to  register namespaces and expand them onto parsed xml. If a namespace is not labeled it is automatically given a label. This update disables that by registering all namespaces, whether they are labeled or not.

This will help with parsing by preventing in-built namespaces from modifying what the developer expects to see in the document.